User-friendly error message on insufficient buffer size
